Use stop words in the app

+++ src/main/java/wordcount/CounterOfWords.java
@@ -2,5 +2,5 @@ package wordcount;
 
 public interface CounterOfWords {
 
-    Integer count(String input);
+    Long count(String input);
 }
